requirements.txt
Appium-Python-Client==0.40
pip==19.0.3
selenium==3.141.0